nuke script tweaked for prod

+++ AeroMembersApp/utilities/nuke.sh
@@ -2,8 +2,11 @@
 
 mv ../migrations/loadItems.py .
 rm ../migrations/0*
-mysql -u plumber -pplumberpass -e 'drop database aeromembersdb'
-mysql -u plumber -pplumberpass -e 'create database aeromembersdb'
+if [ $PRODUCTION != true ]
+then
+	mysql -u plumber -pplumberpass -e 'drop database aeromembersdb'
+	mysql -u plumber -pplumberpass -e 'create database aeromembersdb'
+fi
 python ../../manage.py makemigrations
 mv loadItems.py ../migrations/
 python ../../manage.py migrate
